feat: update OrderHeaders and OrdersHeaderDetails
Aggiornato il metodo CreateSalesOrderHeader, affinchè mappi correttamente anche la tabella dettaglio ordini all'invio di un ordine

+++ BikeVille/Entity/BuildVersion.cs
@@ -11,7 +11,7 @@ public partial class BuildVersion
     /// <summary>
     /// Primary key for BuildVersion records.
     /// </summary>
-    public byte SystemInformationId { get; set; }
+    public int SystemInformationId { get; set; }
 
     /// <summary>
     /// Version number of the database in 9.yy.mm.dd.00 format.
